split and test
